requirements.txt
choices==0.1
dj-database-url==0.5.0
Django==2.0.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-model-utils==3.1.2
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.11.3
pdfkit==0.6.1
Pillow==5.2.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyPDF2==1.26.0
pytz==2018.5
reportlab==3.5.6
six==1.11.0
webencodings==0.5.1
xhtml2pdf==0.2.2
